Move OWL primitives to primitives.py
